\name{missing_recode}
\alias{missing_recode}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function to recode missing data in SNP genotypic data for linkage mapping studies
}
\description{
The function recodes missing data from SNP genotyping to user specified characters. For example, for the data to be used in several linkage mapping softwares, the missing data needs to be recoded to -1,.,NA or other various things. This function automates the process and the user can specify what to call the missing data.
}
\usage{
missing_recode(gendat.mat, missing.data.recode, missing.data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gendat.mat}{
This is a genotypic data matrix with SNP markers in rows and individuals in columns. The first column should contain the marker names while the first row should contain individual IDs. The second column should contain the genotypic information of one of the parent while the third column should contain the genotypic call for each marker of another parent. Ideally in plant breeding linkage mapping studies, population is generated by crossing two distinct individuals and the progenies are used for mapping.
}
  \item{missing.data.recode}{
This object specifies what the missing data should be coded to. For example, NA, 0 or . inside.
}
  \item{missing.data}{
This is what the missing data is reported as before recoding. It could very well be NA or 'failed' as in chip genotyping to indicate failure to make a genotypic call.
}
}
\details{
If necessary, more details than the description above.
}
\value{
This function returns a matrix of genotypic data with missing data recoded, as specified by the user.
}
\references{
For more description using example datasets, refer to GitHub "https://github.com/adhik009" or "aggieanil.wordpress.com".
}
\author{
Anil Adhikari, PhD student, Soil and Crop Science Department, Texas A&M University, College Station, TX.
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (gendat.mat, missing.data.recode, missing.data)
{
    for (i in 1:nrow(gendat.mat)) {
        for (j in 4:ncol(gendat.mat)) {
            if (gendat.mat[i, j] == missing.data)
                gendat.mat[i, j] <- missing.data.recode
        }
    }
    return(gendat.mat)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
